v5

+++ c/bartlett.c
@@ -19,10 +19,10 @@ int bartlett_s (float *Y, const size_t L, const size_t norm)
 {
     if (norm>3u) { fprintf(stderr,"error in bartlett_s: norm must be in {0,1,2,3}\n"); return 1; }
 
-    const float p = 1.0f/(L-(L%2u)*(L/2u)-1u);
+    const float p = 1.0f/(float)(L-(L%2u)*(L/2u)-1u);
     // struct timespec tic, toc; clock_gettime(CLOCK_REALTIME,&tic);
 
-    for (size_t l=0u; l<L/2u; ++l, ++Y) { *Y = (l*(2u-L%2u))*p; }
+    for (size_t l=0u; l<L/2u; ++l, ++Y) { *Y = (float)(l*(2u-L%2u))*p; }
     if (L%2u) { *Y++ = 1.0f; }
     for (size_t l=0u; l<L/2u; ++l, ++Y) { *Y = *(Y-(2u*l+1u+L%2u)); }
 
@@ -45,9 +45,9 @@ int bartlett_d (double *Y, const size_t L, const size_t norm)
 {
     if (norm>3u) { fprintf(stderr,"error in bartlett_d: norm must be in {0,1,2,3}\n"); return 1; }
     
-    const double p = 1.0/(L-(L%2u)*(L/2u)-1u);
+    const double p = 1.0/(double)(L-(L%2u)*(L/2u)-1u);
 
-    for (size_t l=0u; l<L/2u; ++l, ++Y) { *Y = (l*(2u-L%2u))*p; }
+    for (size_t l=0u; l<L/2u; ++l, ++Y) { *Y = (double)(l*(2u-L%2u))*p; }
     if (L%2u) { *Y++ = 1.0; }
     for (size_t l=0u; l<L/2u; ++l, ++Y) { *Y = *(Y-(2u*l+1u+L%2u)); }
 
@@ -68,9 +68,9 @@ int bartlett_c (float *Y, const size_t L, const size_t norm)
 {
     if (norm>3u) { fprintf(stderr,"error in bartlett_c: norm must be in {0,1,2,3}\n"); return 1; }
 
-    const float p = 1.0f/(L-(L%2u)*(L/2u)-1u);
+    const float p = 1.0f/(float)(L-(L%2u)*(L/2u)-1u);
 
-    for (size_t l=0u; l<L/2u; ++l, ++Y) { *Y = (l*(2u-L%2u))*p; *++Y = 0.0f; }
+    for (size_t l=0u; l<L/2u; ++l, ++Y) { *Y = (float)(l*(2u-L%2u))*p; *++Y = 0.0f; }
     if (L%2u) { *Y++ = 1.0f; *Y++ = 0.0f; }
     for (size_t l=0u; l<L/2u; ++l, ++Y) { *Y = *(Y-(4u*l+2u+2u*(L%2u))); *++Y = 0.0f; }
 
@@ -91,9 +91,9 @@ int bartlett_z (double *Y, const size_t L, const size_t norm)
 {
     if (norm>3u) { fprintf(stderr,"error in bartlett_z: norm must be in {0,1,2,3}\n"); return 1; }
 
-    const double p = 1.0/(L-(L%2u)*(L/2u)-1u);
+    const double p = 1.0/(double)(L-(L%2u)*(L/2u)-1u);
     
-    for (size_t l=0u; l<L/2u; ++l, ++Y) { *Y = (l*(2u-L%2u))*p; *++Y = 0.0; }
+    for (size_t l=0u; l<L/2u; ++l, ++Y) { *Y = (double)(l*(2u-L%2u))*p; *++Y = 0.0; }
     if (L%2u) { *Y++ = 1.0; *Y++ = 0.0; }
     for (size_t l=0u; l<L/2u; ++l, ++Y) { *Y = *(Y-(4u*l+2u+2u*(L%2u))); *++Y = 0.0; }
