+++ c/delta_impulse.c
@@ -16,9 +16,9 @@ extern "C" {
 #endif
 
 int delta_impulse_s (float *Y, const size_t N, const size_t tau, const float amp);
-int delta_impulse_d (double *Y, const size_t N, const size_t tau, const float amp);
+int delta_impulse_d (double *Y, const size_t N, const size_t tau, const double amp);
 int delta_impulse_c (float *Y, const size_t N, const size_t tau, const float amp);
-int delta_impulse_z (double *Y, const size_t N, const size_t tau, const float amp);
+int delta_impulse_z (double *Y, const size_t N, const size_t tau, const double amp);
 
 
 int delta_impulse_s (float *Y, const size_t N, const size_t tau, const float amp)
@@ -27,19 +27,19 @@ int delta_impulse_s (float *Y, const size_t N, const size_t tau, const float amp
 
     for (size_t n=0u; n<tau; ++n, ++Y) { *Y = 0.0f; }
     *Y++ = amp;
-    for (size_t n=tau+1; n<N; ++n, ++Y) { *Y = 0.0f; }
+    for (size_t n=tau+1u; n<N; ++n, ++Y) { *Y = 0.0f; }
 
     return 0;
 }
 
 
-int delta_impulse_d (double *Y, const size_t N, const size_t tau, const float amp)
+int delta_impulse_d (double *Y, const size_t N, const size_t tau, const double amp)
 {
     if (tau>=N) { fprintf(stderr, "error in delta_impulse_d: tau (delay) must be less than N\n"); return 1; }
 
     for (size_t n=0u; n<tau; ++n, ++Y) { *Y = 0.0; }
     *Y++ = amp;
-    for (size_t n=tau+1; n<N; ++n, ++Y) { *Y = 0.0; }
+    for (size_t n=tau+1u; n<N; ++n, ++Y) { *Y = 0.0; }
 
     return 0;
 }
@@ -51,19 +51,19 @@ int delta_impulse_c (float *Y, const size_t N, const size_t tau, const float amp
 
     for (size_t n=0u; n<2u*tau; ++n, ++Y) { *Y = 0.0f; }
     *Y++ = amp;
-    for (size_t n=2u*tau+1; n<2u*N; ++n, ++Y) { *Y = 0.0f; }
+    for (size_t n=2u*tau+1u; n<2u*N; ++n, ++Y) { *Y = 0.0f; }
 
     return 0;
 }
 
 
-int delta_impulse_z (double *Y, const size_t N, const size_t tau, const float amp)
+int delta_impulse_z (double *Y, const size_t N, const size_t tau, const double amp)
 {
     if (tau>=N) { fprintf(stderr, "error in delta_impulse_z: tau (delay) must be less than N\n"); return 1; }
 
     for (size_t n=0u; n<2u*tau; ++n, ++Y) { *Y = 0.0; }
     *Y++ = amp;
-    for (size_t n=2u*tau+1; n<2u*N; ++n, ++Y) { *Y = 0.0; }
+    for (size_t n=2u*tau+1u; n<2u*N; ++n, ++Y) { *Y = 0.0; }
 
     return 0;
 }
